fix: -Wsign-compare in the Notebook tab control, 14 -> 0 (132 of 206)
First of the GUI files. Notebook.hpp 9 -> 0, Notebook.cpp 5 -> 0.

Notebook.hpp is a header, so its warnings are reported by every .cpp that includes
it - CreateMMUTiledCanvas was re-reporting all 9 on its own. Fixing the header
clears them everywhere at once.

One is a real bug, in DoRemovePage(size_t page):

    for (int i = 0; i < btidx_to_tabpage.size(); i++)
        if (btidx_to_tabpage[i] > page)
            btidx_to_tabpage[i]--;
        else if (btidx_to_tabpage[i] == page)
            btidx_to_tabpage[i] = -1;

btidx_to_tabpage is a vector<int16_t> in which -1 means "this button has no tab
page". page is a size_t, so the int16_t element was promoted to size_t: the -1
sentinel became SIZE_MAX, compared greater than every real page, and was
decremented to -2 - on every page removal, so it drifts further negative each
time. The loop immediately above already wrote int16_t(page) for its equality
test, which is the same idiom this now uses.

Checked in Python: for real page ids (>= 0) the signed and unsigned forms agree on
both the > and == tests everywhere; they differ only for the -1 sentinel, which is
exactly the case being fixed. Downstream code tests these values with `>= 0`, so a
-2 still reads as "no page" and the drift is latent rather than visibly broken -
but it is wrong and it is what the warning was pointing at.

The rest are index loops over size_t bounds (InsertPage/InsertBtPage/RemoveBtPage
and the four spacer scans in ButtonsListCtrl) plus two comparisons against an
int16_t that is guarded >= 0 at the point of use, where the cast is exact:
SetBtSelection's page-hiding loop and EmitEventSelChanged's page-count bound.

Verified: Notebook.cpp compiles with -Wsign-compare at 0 for both the .cpp and the
header (tu_check --headers), and no other flag's count moved. The application
builds and runs - launched and left up 20 s without exiting, which constructs the
main frame and therefore the settings tab bar, so the InsertBtPage paths execute.

NOT verified at runtime: DoRemovePage itself. A plain launch only adds tabs; pages
are removed when the tab set changes (e.g. switching printer technology), and
driving that needs GUI automation this box does not have set up. The sentinel fix
rests on the Python check and on matching the sibling loop's existing idiom.

The slice panel is no use for any of this - CLI slicing never enters GUI code.

+++ src/slic3r/GUI/Notebook.cpp
@@ -259,7 +259,7 @@ bool ButtonsListCtrl::InsertPage(size_t n, const wxString& text, bool bSelect/* 
     m_pageButtons.insert(m_pageButtons.begin() + n, btn);
     m_spacers.insert(m_spacers.begin() + n, false);
     size_t idx = n;
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         if (m_spacers[i]) idx++;
     }
     m_buttons_sizer->Insert(idx, new wxSizerItem(btn));
@@ -276,7 +276,7 @@ bool ButtonsListCtrl::InsertSpacer(size_t n, int size)
     }
     m_spacers[n] = true;
     size_t idx = n;
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         if (m_spacers[i]) idx++;
     }
     m_buttons_sizer->Insert(idx, size, 1);
@@ -303,7 +303,7 @@ void ButtonsListCtrl::RemovePage(size_t n)
         m_focused_button = nullptr;
     m_pageButtons.erase(m_pageButtons.begin() + n);
     size_t idx = n;
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         if (m_spacers[i]) idx++;
     }
     if (m_spacers[n])
@@ -319,7 +319,7 @@ void ButtonsListCtrl::RemoveSpacer(size_t n)
 {
     if (m_spacers[n]) {
         size_t idx = n;
-        for (int i = 0; i < n; i++) {
+        for (size_t i = 0; i < n; i++) {
             if (m_spacers[i]) idx++;
         }
         m_buttons_sizer->Remove(idx);
@@ -364,7 +364,7 @@ ScalableButton* ButtonsListCtrl::GetPageButton(size_t n)
 void Notebook::EmitEventSelChanged(int16_t new_sel) {
 
     //emit event for changed tab
-    if (new_sel >=0 && GetBtnsListCtrl() && this->GetPageCount() > new_sel) {
+    if (new_sel >=0 && GetBtnsListCtrl() && this->GetPageCount() > size_t(new_sel)) {
         ScalableButton* btn = GetBtnsListCtrl()->GetPageButton(new_sel);
         if (btn) {
             wxCommandEvent* evt = new wxCommandEvent(wxCUSTOMEVT_NOTEBOOK_BT_PRESSED);
